+++ app/decorators/auth-route.ts
